refactor(telegram): extract lifecycle into TelegramLifecycleMixin (adapter god-file slice A2)
(cherry picked from commit 634b7dc31ad0f2fbbbbf750b8745611db713265a)
